[ORGA] Refactor Layout::Topbar component
