+++ orga/app/components/layout/topbar.hbs
@@ -1,0 +1,8 @@
+<div class="topbar">
+  <div class="topbar__organization-credit-info hide-on-mobile">
+    <Layout::OrganizationCreditInfo />
+  </div>
+  <div class="topbar__user-logged-menu">
+    <Layout::UserLoggedMenu />
+  </div>
+</div>
