+++ orga/app/styles/app.scss
@@ -28,7 +28,10 @@
 @import "components/dropdown";
 @import "components/indicator-card";
 @import "components/join-request-form";
+@import "components/layout/organization-credit-info";
 @import "components/layout/sidebar";
+@import "components/layout/topbar";
+@import "components/layout/user-logged-menu";
 @import "components/pagination-control";
 @import "components/participant/tabs";
 @import "components/previous-page-button";
@@ -39,10 +42,8 @@
 @import "components/select-input";
 @import "components/team";
 @import "components/login-or-register";
-@import "components/user-logged-menu";
 @import "components/manage-authentication-method-modal";
 @import "components/modal";
-@import "components/organization-credit-info";
 @import "pages/login";
 @import "pages/join";
 @import "pages/join-request";
