+++ orga/app/styles/components/layout/organization-credit-info.scss
